moved classes to models directory; optional fields no longer have default values
